properly move deps
