+++ projects/orquestra-sdk/src/orquestra/sdk/_client/mlflow/_connection_utils.py
@@ -10,11 +10,11 @@ import warnings
 from pathlib import Path
 from typing import Optional, Tuple
 
+from orquestra.workflow_shared.env import PASSPORT_FILE_ENV
 from orquestra.workflow_shared.exceptions import (
     ConfigNameNotFoundError,
     RuntimeConfigError,
 )
-from orquestra.workflow_shared.env import PASSPORT_FILE_ENV
 from requests import Response, Session
 
 from orquestra import sdk
